🚚 api: get-certification-attestation (first route) : move
usecase to certification course src
