+++ api/src/certification/shared/domain/usecases/index.js
@@ -8,6 +8,7 @@ import * as certificationCenterRepository from '../../../shared/infrastructure/r
 import * as certificationCandidateRepository from '../../../shared/infrastructure/repositories/certification-candidate-repository.js';
 import * as certificationChallengeLiveAlertRepository from '../../../session/infrastructure/repositories/certification-challenge-live-alert-repository.js';
 import * as certificationIssueReportRepository from '../../../shared/infrastructure/repositories/certification-issue-report-repository.js';
+import * as certificateRepository from '../../../../../lib/infrastructure/repositories/certificate-repository.js';
 import * as challengeRepository from '../../../../shared/infrastructure/repositories/challenge-repository.js';
 import * as certificationCpfCountryRepository from '../../../shared/infrastructure/repositories/certification-cpf-country-repository.js';
 import * as certificationCpfCityRepository from '../../../shared/infrastructure/repositories/certification-cpf-city-repository.js';
@@ -38,6 +39,7 @@ const dependencies = {
   certificationChallengeLiveAlertRepository,
   certificationCpfService,
   certificationIssueReportRepository,
+  certificateRepository,
   challengeRepository,
   certificationCpfCityRepository,
   certificationCpfCountryRepository,
@@ -69,6 +71,9 @@ const usecasesWithoutInjectedDependencies = {
   })),
   ...(await importNamedExportsFromDirectory({
     path: join(path, '../../../flash-certification/domain/usecases/'),
+  })),
+  ...(await importNamedExportsFromDirectory({
+    path: join(path, '../../../course/domain/usecases/'),
   })),
 };
